+++ src/index.rs
@@ -36,7 +36,7 @@
 //! ```
 
 use crate::Token;
-use alloc::string::String;
+use alloc::{string::String, vec::Vec};
 use core::{fmt, num::ParseIntError, str::FromStr};
 
 /// Represents an abstract index into an array.
@@ -166,7 +166,21 @@ impl FromStr for Index {
         } else if s.starts_with('0') && s != "0" {
             Err(ParseIndexError::LeadingZeros)
         } else {
-            Ok(s.parse::<usize>().map(Index::Num)?)
+            let idx = s.parse::<usize>().map(Index::Num)?;
+            if s.chars().all(|c| c.is_ascii_digit()) {
+                Ok(idx)
+            } else {
+                // this comes up with the `+` sign which is valid for
+                // representing a `usize` but not allowed in RFC 6901 array
+                // indices
+                let mut invalid: Vec<_> = s.chars().filter(|c| !c.is_ascii_digit()).collect();
+                // remove duplicate characters
+                invalid.sort_unstable();
+                invalid.dedup();
+                Err(ParseIndexError::InvalidCharacters(
+                    invalid.into_iter().collect(),
+                ))
+            }
         }
     }
 }
@@ -267,6 +281,8 @@ pub enum ParseIndexError {
     InvalidInteger(ParseIntError),
     /// The Token contains leading zeros.
     LeadingZeros,
+    /// The Token contains non-digit characters.
+    InvalidCharacters(String),
 }
 
 impl From<ParseIntError> for ParseIndexError {
@@ -285,6 +301,11 @@ impl fmt::Display for ParseIndexError {
                 f,
                 "token contained leading zeros, which are disallowed by RFC 6901"
             ),
+            ParseIndexError::InvalidCharacters(chars) => write!(
+                f,
+                "token contains non-digit character(s) '{chars}', \
+                which are disallowed by RFC 6901",
+            ),
         }
     }
 }
@@ -294,7 +315,7 @@ impl std::error::Error for ParseIndexError {
     fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
         match self {
             ParseIndexError::InvalidInteger(source) => Some(source),
-            ParseIndexError::LeadingZeros => None,
+            ParseIndexError::LeadingZeros | ParseIndexError::InvalidCharacters(_) => None,
         }
     }
 }
@@ -412,6 +433,11 @@ mod tests {
         assert_eq!(
             ParseIndexError::LeadingZeros.to_string(),
             "token contained leading zeros, which are disallowed by RFC 6901"
+        );
+        assert_eq!(
+            ParseIndexError::InvalidCharacters("+@".into()).to_string(),
+            "token contains non-digit character(s) '+@', \
+                which are disallowed by RFC 6901"
         );
     }