+++ CHANGELOG.md
@@ -14,11 +14,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Adds unsafe associated methods `Pointer::new_unchecked` and `PointerBuf::new_unchecked` for
     external zero-cost construction.
 -   Adds `Pointer::starts_with` and `Pointer::ends_with` for prefix and suffix matching.
+-   Adds new `ParseIndexError` variant to express the presence non-digit characters in the token.
+-   Adds `Token::is_next` for checking if a token represents the `-` character.
 
 ### Changed
 
 -   Changed signature of `PathBuf::parse` to avoid requiring allocation.
 -   Bumps minimum Rust version to 1.79.
+-   `Pointer::get` now accepts ranges and can produce `Pointer` segments as output (similar to
+    `slice::get`).
+
+### Fixed
+
+-   Make validation of array indices conform to RFC 6901 in the presence of non-digit characters.
 
 ## [0.6.2] 2024-09-30